Refactor to move the Editor to CreatorUtilities (#102)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change moves the Editor abstract class to be CreatorUtilities at the same level as the Creator. This is inline with FAST Tooling naming standards and is a step in the removal of the site-utilities folder.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check that the site still works and looks as expected from the generated link.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Remove other components from the site-utilities
