+++ app/site-utilities/components/index.ts
@@ -1,6 +1,5 @@
 export * from "./dimension";
 export * from "./direction-switch";
-export * from "./editor";
 export * from "./logo";
 export * from "./theme-selector";
 export * from "./transparency-toggle";
